+++ src/components/panels/InfoPanel.tsx
@@ -1,0 +1,31 @@
+import { Entry, Place, Product } from "../../types";
+import FountainPanel from "./FountainPanel";
+import GastroPanel from "./GastroPanel";
+import KioskPanel from "./KioskPanel";
+
+export default function InfoPanel({
+  activePlace,
+  activeEntries,
+  products,
+  unsetActivePlace,
+}: {
+  activePlace: Place;
+  activeEntries: Entry[] | undefined;
+  products: Map<number, Product>;
+  unsetActivePlace: () => void;
+}) {
+  if (activePlace.source === "drop") {
+    return <FountainPanel activePlace={activePlace} unsetActivePlace={unsetActivePlace} />;
+  }
+  if (activePlace.source === "bag") {
+    return <KioskPanel activePlace={activePlace} unsetActivePlace={unsetActivePlace} />;
+  }
+  return (
+    <GastroPanel
+      activePlace={activePlace}
+      activeEntries={activeEntries}
+      products={products}
+      unsetActivePlace={unsetActivePlace}
+    />
+  );
+}
