+++ src/components/panels/KioskPanel.tsx
@@ -1,0 +1,22 @@
+import { Place } from "../../types";
+import { CloseButton } from "../Buttons";
+import { formatHours, formatPlaceType, googlifyAddress } from "./format";
+
+export default function KioskPanel({
+  activePlace,
+  unsetActivePlace,
+}: {
+  activePlace: Place;
+  unsetActivePlace: () => void;
+}) {
+  return (
+    <div className="info-panel">
+      <CloseButton onClick={unsetActivePlace} />
+      <h3>{activePlace.placeName}</h3>
+      {formatPlaceType(activePlace.placeType)}
+      <div className="info-divider" />
+      {googlifyAddress(activePlace.address)}
+      {formatHours(activePlace.note)}
+    </div>
+  );
+}
